+++ src/main/java/org/opentripplanner/api/common/RoutingResource.java
@@ -425,7 +425,14 @@ public abstract class RoutingResource {
     @QueryParam("serviceDayLookout")
     private Integer serviceDayLookout;
 
-    /* 
+    /**
+     * Set the method of sorting itineraries in the response. Right now, the only supported value is "duration";
+     * otherwise it uses default sorting. More sorting methods may be added in the future.
+     */
+    @QueryParam("pathComparator")
+    private String pathComparator;
+
+    /*
      * somewhat ugly bug fix: the graphService is only needed here for fetching per-graph time zones. 
      * this should ideally be done when setting the routing context, but at present departure/
      * arrival time is stored in the request as an epoch time with the TZ already resolved, and other
@@ -697,6 +704,9 @@ public abstract class RoutingResource {
 
         if (serviceDayLookout != null)
             request.setServiceDayLookout(serviceDayLookout);
+
+        if (pathComparator != null)
+            request.pathComparator = pathComparator;
 
         //getLocale function returns defaultLocale if locale is null
         request.locale = ResourceBundleSingleton.INSTANCE.getLocale(locale);